Move code to respective modules.
