+++ farm/tests/src/lib.rs
@@ -1,4 +1,6 @@
+#[cfg(test)]
 mod farm;
+#[cfg(test)]
 mod psp22;
 #[cfg(test)]
 mod tests;
