+++ farm/tests/src/tests.rs
@@ -1,27 +1,27 @@
 use crate::*;
 
-use farm::{FarmDetails, FarmError, FarmT, PSP22Error};
+use farm::{self, FarmDetails, FarmError, PSP22Error};
+use psp22;
 use utils::*;
 
 use drink::{runtime::MinimalRuntime, session::Session};
-use ink_wrapper_types::Connection;
 
 #[test]
 fn farm_start() {
     let mut session: Session<MinimalRuntime> = Session::new().expect("Init new Session");
 
-    let ice = setup_psp22(&mut session, ICE.to_string(), ICE.to_string(), BOB);
-    let wood = setup_psp22(&mut session, WOOD.to_string(), WOOD.to_string(), BOB);
-    let sand = setup_psp22(&mut session, SAND.to_string(), SAND.to_string(), BOB);
+    let ice = psp22::setup(&mut session, ICE.to_string(), ICE.to_string(), BOB);
+    let wood = psp22::setup(&mut session, WOOD.to_string(), WOOD.to_string(), BOB);
+    let sand = psp22::setup(&mut session, SAND.to_string(), SAND.to_string(), BOB);
 
-    let farm = setup_farm(
+    let farm = farm::setup(
         &mut session,
         ice.into(),
         vec![wood.into(), sand.into()],
         BOB,
     );
 
-    let farm_details: FarmDetails = get_farm_details(&mut session, &farm);
+    let farm_details: FarmDetails = farm::get_farm_details(&mut session, &farm);
 
     let expected_details = FarmDetails {
         pool_id: ice.into(),
@@ -38,30 +38,34 @@ fn farm_start() {
     set_timestamp(&mut session, now);
     let farm_start = now + 100;
     let farm_end = farm_start + 100;
+    let rewards_amount = 100;
 
-    let call_result = session
-        .query(farm.owner_start_new_farm(farm_start, farm_end, vec![100, 100]))
-        .unwrap()
-        .result
-        .unwrap();
+    let start_result = farm::start(
+        &mut session,
+        &farm,
+        farm_start,
+        farm_end,
+        vec![rewards_amount, rewards_amount],
+        BOB,
+    );
 
     let insufficient_allowance = FarmError::PSP22Error(PSP22Error::InsufficientAllowance());
 
-    assert!(
-        call_result == Err(insufficient_allowance),
+    assert_eq!(
+        start_result,
+        Err(insufficient_allowance),
         "Caller hasn't increased allowance to spend reward tokens for the farm"
     );
 
-    let rewards_amount = 100;
-    increase_allowance(&mut session, wood.into(), farm.into(), rewards_amount, BOB).unwrap();
-    increase_allowance(&mut session, sand.into(), farm.into(), rewards_amount, BOB).unwrap();
+    psp22::increase_allowance(&mut session, wood.into(), farm.into(), rewards_amount, BOB).unwrap();
+    psp22::increase_allowance(&mut session, sand.into(), farm.into(), rewards_amount, BOB).unwrap();
 
-    let bob_wood_balance_before = balance_of(&mut session, wood.into(), bob());
-    let bob_sand_balance_before = balance_of(&mut session, sand.into(), bob());
-    let farm_wood_balance_before = balance_of(&mut session, wood.into(), farm.into());
-    let farm_sand_balance_before = balance_of(&mut session, sand.into(), farm.into());
+    let bob_wood_balance_before = psp22::balance_of(&mut session, wood.into(), bob());
+    let bob_sand_balance_before = psp22::balance_of(&mut session, sand.into(), bob());
+    let farm_wood_balance_before = psp22::balance_of(&mut session, wood.into(), farm.into());
+    let farm_sand_balance_before = psp22::balance_of(&mut session, sand.into(), farm.into());
 
-    let call_result = setup_farm_start(
+    let call_result = farm::start(
         &mut session,
         &farm,
         farm_start,
@@ -80,12 +84,12 @@ fn farm_start() {
         end: farm_end,
     };
 
-    let farm_details = get_farm_details(&mut session, &farm);
+    let farm_details = farm::get_farm_details(&mut session, &farm);
 
     assert!(farm_details == expected_details);
 
-    let bob_wood_balance_after = balance_of(&mut session, wood.into(), bob());
-    let bob_sand_balance_after = balance_of(&mut session, sand.into(), bob());
+    let bob_wood_balance_after = psp22::balance_of(&mut session, wood.into(), bob());
+    let bob_sand_balance_after = psp22::balance_of(&mut session, sand.into(), bob());
     assert!(
         bob_wood_balance_after == bob_wood_balance_before - rewards_amount,
         "Farm start must deduct rewards from the caller"
@@ -94,8 +98,8 @@ fn farm_start() {
         bob_sand_balance_after == bob_sand_balance_before - rewards_amount,
         "Farm start must deduct rewards from the caller"
     );
-    let farm_wood_balance_after = balance_of(&mut session, wood.into(), farm.into());
-    let farm_sand_balance_after = balance_of(&mut session, sand.into(), farm.into());
+    let farm_wood_balance_after = psp22::balance_of(&mut session, wood.into(), farm.into());
+    let farm_sand_balance_after = psp22::balance_of(&mut session, sand.into(), farm.into());
     assert!(
         farm_wood_balance_after == farm_wood_balance_before + rewards_amount,
         "Farm start must transfer rewards to the farm"
